[TWINFACES-233] used ComboboxFormField for the dropdown logic
