+++ src/features/datalist/index.ts
@@ -1,2 +1,1 @@
 export * from "./context-provider";
-export * from "./select-field";
