+++ src/features/permission/index.ts
@@ -1,4 +1,3 @@
 "use client";
 
 export * from "./context-provider";
-export * from "./select-field";
