--- src/features/permission/select-field/index.ts
@@ -1,1 +1,0 @@
-export * from "./permisson-select-field";
